+++ HISTORY.rst
@@ -3,6 +3,7 @@
 History
 =======
 
+- tooltips added to demos launcher
 - added get_username_password
 
 0.4.0